# docs/swagger.yaml
basePath: /
definitions:
...
  model.Usage:
    properties:
      bytes:
        type: integer
      bytesPerDay:
        type: number
      deviceId:
        type: string
      exportId:
        type: string
      updatedAt:
        type: string
    type: object
...
info:
  contact: {}
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Timescale Wrapper API
  version: "0.1"
paths:
...
  /usage/devices:
    get:
      consumes:
      - application/json
      parameters:
      - description: device_ids
        in: body
        name: device_ids
        required: true
        schema:
          items:
            type: string
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: usage
          schema:
            items:
              $ref: '#/definitions/model.Usage'
            type: array
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: Device Usage
  /usage/exports:
    get:
      consumes:
      - application/json
      parameters:
      - description: export_ids
        in: body
        name: export_ids
        required: true
        schema:
          items:
            type: string
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: usage
          schema:
            items:
              $ref: '#/definitions/model.Usage'
            type: array
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: Export Usage
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"